Map Square Density Method Switch (WIP 3)
Renamed the script file and fixed several errors with the code, although one still remains.
